Add new job to dump public db to release db
Does a dump of main cbioportal database and then puts them in a temp
database of the release. This way the existing beta instance can
function without problems. Only once the entire temp db is ready, the
release db is removed and temp db is renamed to the new release db. This
should only be run when cbioportal is not in use, because it's pretty
heavy on the RDS. We should prolly set something up to push the db dumps
to S3 or something. Then we could use those dumps from S3 instead of
targeting the production db. Note that one still needs to manually give
the database grant acces, once it's initialized.