Add demo for HTAN using mira for single cell vis
This shows how one might use mira for single cell visualization. It is
not sure yet whether this is the tool we want to use for HTAN, but it
demonstrates feasibility. Note that the README does not indicate how to
load data, only how to run the system on kubernetes. For more info on
data loading see:
https://github.com/cBioPortal/cbioportal/issues/7323