Genome Nexus v1.1.8 + use ZGC garbage collector
We had some Genome Nexus downtime during last week's webinar potentially
because of memory issues and GC pauses. The hypothesis is that the
increased memory would kick in the GC, which in turn would cause the
liveness/readiness probes to fail and cause the system to go down. For
cBioPortal we have been using the shenandoah gc with some success. It
does more active garbage collection at a cost of CPU. For Genome Nexus
we switched to the openjdk 13 image from Oracle since v1.1.8. This
doesn't support Shenandoah out of the box which is from Redhat, but it
does Oracle's ZGC garbage collector which is supposed to be similar (a
low pause time garbage collector). For a more in-depth comparison see:
https://blog.plan99.net/modern-garbage-collection-part-2-1c88847abcfd